Rework etcd-launcher to allow multiple commands

Temporarily disable etcd-running check
