+++ cmd/etcd-launcher/cmd_health.go
@@ -1,0 +1,1 @@
+package main
