+++ cmd/etcd-launcher/cmd_maintenance.go
@@ -1,0 +1,1 @@
+package main
