+++ pkg/resources/etcd/etcdrunning/etcdrunning.go
@@ -17,13 +17,7 @@ limitations under the License.
 package etcdrunning
 
 import (
-	"fmt"
-	"strings"
-
 	kubermaticv1 "k8c.io/kubermatic/v2/pkg/apis/kubermatic/v1"
-	"k8c.io/kubermatic/v2/pkg/resources"
-	"k8c.io/kubermatic/v2/pkg/resources/etcd"
-	"k8c.io/kubermatic/v2/pkg/resources/registry"
 
 	corev1 "k8s.io/api/core/v1"
 )
@@ -34,21 +28,23 @@ type etcdRunningData interface {
 }
 
 func Container(etcdEndpoints []string, data etcdRunningData) corev1.Container {
-	return corev1.Container{
-		Name:  "etcd-running",
-		Image: registry.Must(data.RewriteImage(resources.RegistryGCR + "/etcd-development/etcd:" + etcd.ImageTag(data.Cluster()))),
-		Command: []string{
-			"/bin/sh",
-			"-ec",
-			// Write a key to etcd. If we have quorum it will succeed. after 100 retries the script return an error
-			fmt.Sprintf("for i in $(seq 1 100); do if ETCDCTL_API=3 /usr/local/bin/etcdctl --cacert=/etc/etcd/pki/client/ca.crt --cert=/etc/etcd/pki/client/apiserver-etcd-client.crt --key=/etc/etcd/pki/client/apiserver-etcd-client.key --dial-timeout=2s --endpoints='%s' put kubermatic/quorum-check something; then echo \"etcd ready\"; exit 0; fi; echo \"waiting for etcd. retry=$i/100\"; sleep 2; done; echo \"error: etcd not ready\"; exit 1;", strings.Join(etcdEndpoints, ",")),
-		},
-		VolumeMounts: []corev1.VolumeMount{
-			{
-				Name:      resources.ApiserverEtcdClientCertificateSecretName,
-				MountPath: "/etc/etcd/pki/client",
-				ReadOnly:  true,
+	return corev1.Container{}
+	/*
+			Name:  "etcd-running",
+			Image: registry.Must(data.RewriteImage(resources.RegistryGCR + "/etcd-development/etcd:" + etcd.ImageTag(data.Cluster()))),
+			Command: []string{
+				"/bin/sh",
+				"-ec",
+				// Write a key to etcd. If we have quorum it will succeed. after 100 retries the script return an error
+				fmt.Sprintf("for i in $(seq 1 100); do if ETCDCTL_API=3 /usr/local/bin/etcdctl --cacert=/etc/etcd/pki/client/ca.crt --cert=/etc/etcd/pki/client/apiserver-etcd-client.crt --key=/etc/etcd/pki/client/apiserver-etcd-client.key --dial-timeout=2s --endpoints='%s' put kubermatic/quorum-check something; then echo \"etcd ready\"; exit 0; fi; echo \"waiting for etcd. retry=$i/100\"; sleep 2; done; echo \"error: etcd not ready\"; exit 1;", strings.Join(etcdEndpoints, ",")),
 			},
-		},
-	}
+			VolumeMounts: []corev1.VolumeMount{
+				{
+					Name:      resources.ApiserverEtcdClientCertificateSecretName,
+					MountPath: "/etc/etcd/pki/client",
+					ReadOnly:  true,
+				},
+			},
+		}
+	*/
 }
